+++ src/main/java/com/api/teacher/TeacherService.java
@@ -1,6 +1,9 @@
 package com.api.teacher;
 
 import com.api.common.error.exceptions.BadRequestException;
+import com.api.common.error.exceptions.NotFoundException;
+import com.api.common.error.exceptions.UnauthorizedException;
+import com.api.login.Auth;
 import com.api.login.LoginService;
 import com.api.teacher.dto.CreateTeacherRequest;
 import com.api.teacher.dto.TeacherDto;
@@ -24,7 +27,10 @@ public class TeacherService {
     }
 
     @Transactional
-    public TeacherDto create(CreateTeacherRequest request) {
+    public TeacherDto create(CreateTeacherRequest request, Auth auth) {
+        if(!auth.getRole().equalsIgnoreCase("admin")){
+            throw new UnauthorizedException("Only admin can create teacher");
+        }
         if (!loginService.isEmailAvailable(request.getEmail())) {
             throw new BadRequestException("User with this email already exists.");
         }
@@ -34,28 +40,28 @@ public class TeacherService {
     }
 
     @Transactional
-    public List<TeacherDto> readAll() {
-       return teacherRepository.findAll()
+    public List<TeacherDto> readAll(Auth auth) {
+        return teacherRepository.findAll()
                .stream()
                .map(teacherMapper::toDto)
                .toList();
     }
 
     @Transactional
-    public TeacherDto read(int teacherID) {
-        if (!teacherRepository.existsByTeacherID(teacherID)) {
-            throw new BadRequestException("Teacher doesn't exist.");
-        }
-
-        Teacher teacher = teacherRepository.getReferenceById(teacherID);
+    public TeacherDto read(int teacherID, Auth auth) {
+        Teacher teacher = teacherRepository.findById(teacherID)
+                .orElseThrow(() -> new NotFoundException("Teacher doesn't exist."));
         return teacherMapper.toDto(teacher);
     }
 
     @Transactional
-    public TeacherDto update(int teacherID, UpdateTeacherRequest request) {
+    public TeacherDto update(int teacherID, UpdateTeacherRequest request,Auth auth) {
         Teacher teacher = teacherRepository.findById(teacherID)
                 .orElseThrow(() -> new BadRequestException("Teacher doesn't exist."));
-
+        if(!auth.getRole().equalsIgnoreCase("admin")
+                && !auth.getEmail().equalsIgnoreCase(teacher.getEmail())){
+            throw new UnauthorizedException("Only admin and self can update teacher");
+        }
         teacherMapper.updateEntity(teacher, request);
         teacherRepository.save(teacher);
 
@@ -63,10 +69,14 @@ public class TeacherService {
     }
 
     @Transactional
-    public void delete(int teacherID) {
+    public void delete(int teacherID,Auth auth) {
         Teacher teacher = teacherRepository.findById(teacherID)
                 .orElseThrow(() -> new BadRequestException("Teacher doesn't exist."));
 
+        if (!auth.getRole().equalsIgnoreCase("admin")
+                || auth.getEmail().equalsIgnoreCase(teacher.getEmail())) {
+            throw new UnauthorizedException("Admin cannot delete themselves");
+        }
         teacherRepository.delete(teacher);
     }
 }
